refactor: extract upload processing logic into dedicated services
Major refactoring to improve separation of concerns and code maintainability:

## New Services Created:
- src/services/upload/single-file-processor.ts - Single file upload business logic
- src/services/upload/multiple-files-processor.ts - Multiple files upload business logic (directory + multiple-files)

## Hook Refactoring:
- useFileUpload hook now focuses purely on React concerns (state, context, event handling)
- Removed ~150 lines of duplicated business logic from the hook
- Hook now routes to appropriate services based on upload mode

## Service Architecture:
- Single file processing: handleFileUpload → processSingleFile → processSingleFileService
- Multiple files processing: handleFileUpload → processMultipleFilesService
- Clean separation between React hook logic and business logic

## Benefits:
- Better testability - services can be unit tested independently
- Improved reusability - services can be used by other components
- Cleaner code - hook is now much more readable and maintainable
- Eliminated code duplication between directory and multiple-files modes
- Better separation of concerns - React vs business logic

## Files Modified:
- src/hooks/use-file-upload.ts - Simplified to focus on React concerns
- src/services/upload/index.ts - Added exports for new services
- src/app/api/storage/edges/route.ts - Updated for new schema structure

This refactoring maintains all existing functionality while significantly improving code organization and maintainability.
