+++ src/hooks/use-file-upload.ts
@@ -4,16 +4,18 @@ import { useState, useRef } from 'react';
 import { useToast } from '@/hooks/use-toast';
 import { useOnboarding } from '@/contexts/onboarding-context';
 import { useSession } from 'next-auth/react';
-import { uploadFile } from '@/services/upload/upload';
+// import { uploadFile } from '@/services/upload/upload'; // Now handled by processSingleFileService
 // Lazy import to avoid eager loading of ICP declarations
 // import { icpUploadService } from '@/services/upload/icp-upload';
 // We'll need to create this context for post-onboarding state
 // import { useVault } from '@/contexts/vault-context';
 // import { isUploadStorageExpired } from '@/hooks/use-upload-storage'; // Not used anymore
 import { useHostingPreferences } from '@/hooks/use-storage-preferences';
-import { verifyIntent } from '@/services/upload/intent';
-import { verifyUpload } from '@/services/upload/verification';
-import { UPLOAD_LIMITS } from '@/config/upload-limits';
+// import { verifyIntent } from '@/services/upload/intent'; // Now handled by services
+// import { verifyUpload } from '@/services/upload/verification'; // Now handled by services
+import { processSingleFile as processSingleFileService } from '@/services/upload/single-file-processor';
+import { processMultipleFiles as processMultipleFilesService } from '@/services/upload/multiple-files-processor';
+// import { UPLOAD_LIMITS } from '@/config/upload-limits'; // Now handled by services
 import type { UseFileUploadProps } from '@/types/upload';
 
 export function useFileUpload({ isOnboarding = false, mode = 'directory', onSuccess, onError }: UseFileUploadProps) {
@@ -93,12 +95,7 @@ export function useFileUpload({ isOnboarding = false, mode = 'directory', onSucc
   // DOM manipulation logic moved to src/lib/file-picker.ts
   // Components should call triggerFileInput(fileInputRef.current, mode) directly
 
-  const checkFileSize = (file: File) => {
-    if (!UPLOAD_LIMITS.isFileSizeValid(file.size)) {
-      console.error('❌ File too large:', UPLOAD_LIMITS.getFileSizeErrorMessage(file.size));
-      throw new Error(UPLOAD_LIMITS.getFileSizeErrorMessage(file.size));
-    }
-  };
+  // File size validation moved to direct UPLOAD_LIMITS usage for consistency
 
   const updateOnboardingContext = (data: { data: { ownerId: string; id: string } }, file: File, url: string) => {
     // console.log("👤 Updating user data with:", {
@@ -134,136 +131,17 @@ export function useFileUpload({ isOnboarding = false, mode = 'directory', onSucc
   };
 
   const processSingleFile = async (file: File, skipSuccess = false, existingUserId?: string) => {
-    console.log(`📁 Processing single file: ${file.name} (${file.size} bytes)`);
-    console.log(`📊 DASHBOARD UPLOAD ANALYSIS:`, {
-      fileName: file.name,
-      fileSize: file.size,
-      fileSizeMB: (file.size / (1024 * 1024)).toFixed(2),
-      fileType: file.type,
+    await processSingleFileService({
+      file,
       isOnboarding,
       mode,
       existingUserId,
-      isLargeFile: file.size / (1024 * 1024) > 4,
-      threshold: '4MB',
+      preferences,
+      onSuccess: skipSuccess ? undefined : onSuccess,
+      onError,
+      updateOnboardingContext,
+      showToast: toast,
     });
-
-    try {
-      checkFileSize(file);
-
-      // Create a temporary URL for preview
-      const url = URL.createObjectURL(file);
-
-      // Use unified upload service with storage preference
-      const userBlobHosting = preferences?.blobHosting; // "s3" | "vercel_blob" | "icp" | "arweave" | "ipfs" | "neon"
-      console.log(`🔍 User blob hosting: ${userBlobHosting}`);
-
-      // For ICP preference, check authentication first
-      if (userBlobHosting === 'icp') {
-        console.log(`🔐 Checking ICP authentication...`);
-        const { icpUploadService } = await import('@/services/upload/icp-upload');
-        const isAuthenticated = await icpUploadService.isAuthenticated();
-        if (!isAuthenticated) {
-          throw new Error('Please connect your Internet Identity to upload to ICP');
-        }
-        console.log(`✅ ICP authentication confirmed`);
-      }
-
-      // Temporary override for testing - force S3 uploads
-      const storageBackend = 's3' as const;
-      console.log('🔧 TEMPORARY OVERRIDE: Forcing S3 uploads for testing');
-      // Original code (commented out for reference):
-      // let storageBackend: 'vercel_blob' | 's3' = 'vercel_blob';
-      // if (userStoragePreference === 's3') {
-      //   storageBackend = 's3';
-      // }
-
-      // Use the unified uploadFile function
-      console.log(`🚀 Calling uploadFile with parameters:`, {
-        fileName: file.name,
-        isOnboarding,
-        existingUserId,
-        mode,
-        storageBackend,
-        userBlobHosting,
-      });
-
-      const uploadResult = await uploadFile(file, isOnboarding, existingUserId, mode, storageBackend, userBlobHosting);
-
-      // Convert to expected format for compatibility
-      // Note: uploadResult.data is an array of memories from the backend
-      const memory = Array.isArray(uploadResult.data) ? uploadResult.data[0] : uploadResult.data;
-
-      // Check if we have a valid memory response
-      if (!memory || !memory.id) {
-        console.error('❌ Invalid upload response:', uploadResult);
-        throw new Error('Upload failed: Invalid response from server');
-      }
-
-      const data = {
-        data: { id: memory.id },
-        results: [
-          {
-            memoryId: memory.id,
-            size: file.size, // Use original file size since assets array might not be available
-            checksum_sha256: null, // Will be filled by verification if available
-          },
-        ],
-        userId: existingUserId || '', // Add userId for compatibility
-      };
-
-      // 3) Verify after upload (best-effort) - only for non-ICP flows
-      // ICP flows handle verification internally
-      const appMemoryId = data?.data?.id;
-      if (appMemoryId && userBlobHosting !== 'icp') {
-        // For non-ICP flows, we still need to get storage info for verification
-        const storageResponse = await verifyIntent({
-          preferred: preferences?.blobHosting,
-          databaseHosting: preferences?.databaseHosting,
-          backendHosting: preferences?.backendHosting,
-        });
-        const storage = storageResponse.uploadStorage;
-        await verifyUpload({
-          appMemoryId,
-          database: storage.database,
-          blob_storage: storage.blob_storage,
-          idem: storage.idem,
-          size: file.size,
-          checksum_sha256: data?.results?.[0]?.checksum_sha256 ?? null,
-          remote_id: data?.results?.[0]?.memoryId ?? appMemoryId,
-        });
-      }
-
-      if (isOnboarding && data) {
-        updateOnboardingContext(
-          { data: { ownerId: data.userId ?? '', id: data.data?.id ?? appMemoryId ?? '' } },
-          file,
-          url
-        );
-      }
-
-      if (!skipSuccess) {
-        onSuccess?.();
-      }
-    } catch (error) {
-      let title = 'Something went wrong';
-      let description = 'Please try uploading again.';
-
-      if (error instanceof Error && error.message.includes('File too large')) {
-        title = 'File too large';
-        description = error.message; // Use the detailed error message from UPLOAD_LIMITS
-      }
-
-      if (error instanceof Error && error.message.includes('intent')) {
-        title = 'Upload not ready';
-        description = error.message;
-      }
-
-      console.error('❌ Upload error:', error);
-
-      toast({ variant: 'destructive', title, description });
-
-      onError?.(error as Error);
-    }
   };
 
   const handleFileUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -274,141 +152,20 @@ export function useFileUpload({ isOnboarding = false, mode = 'directory', onSucc
       const files = event.target.files;
       if (!files) return;
 
-      // Check file count limit
-      if (!UPLOAD_LIMITS.isFileCountValid(files.length)) {
-        toast({
-          variant: 'destructive',
-          title: 'Too many files',
-          description: UPLOAD_LIMITS.getFileCountErrorMessage(files.length),
-        });
-        return;
-      }
-
-      // Check total size limit
-      const totalSize = Array.from(files).reduce((sum, file) => sum + file.size, 0);
-      if (!UPLOAD_LIMITS.isTotalSizeValid(totalSize)) {
-        toast({
-          variant: 'destructive',
-          title: 'Upload too large',
-          description: UPLOAD_LIMITS.getTotalSizeErrorMessage(totalSize),
-        });
-        return;
-      }
-
       setIsLoading(true);
-
       try {
-        // 1) Request upload storage (MVP mock)
-        const storageResponse = await verifyIntent({
-          preferred: preferences?.blobHosting,
-          databaseHosting: preferences?.databaseHosting,
-          backendHosting: preferences?.backendHosting,
+        await processMultipleFilesService({
+          files: Array.from(files),
+          mode: 'directory',
+          isOnboarding,
+          preferences,
+          onSuccess,
+          onError,
+          updateUserData,
+          setCurrentStep,
+          session,
+          showToast: toast,
         });
-        const storage = storageResponse.uploadStorage;
-
-        let data:
-          | {
-              results?: Array<{
-                memoryId: string;
-                size?: number;
-                checksum_sha256?: string | null;
-                name?: string;
-                type?: string;
-              }>;
-              userId?: string;
-              successfulUploads?: number;
-            }
-          | undefined;
-
-        // 2) Route to appropriate upload service based on chosen storage
-        if (storage.blob_storage === 'icp') {
-          // Pre-check Internet Identity authentication before attempting ICP upload
-          const { icpUploadService } = await import('@/services/upload/icp-upload');
-          const isAuthenticated = await icpUploadService.isAuthenticated();
-          if (!isAuthenticated) {
-            throw new Error('Please connect your Internet Identity to upload to ICP');
-          }
-
-          const icpResults = await icpUploadService.uploadFolder(Array.from(files), storage, () => {});
-
-          // Convert ICP results to expected format
-          data = {
-            results: icpResults.map((result, index) => ({
-              memoryId: result.memoryId,
-              size: result.size,
-              name: files[index].name,
-              type: files[index].type,
-              checksum_sha256: result.checksum_sha256,
-            })),
-          };
-        } else {
-          // Upload folder using unified POST /api/memories endpoint
-          const formData = new FormData();
-          Array.from(files).forEach(file => {
-            formData.append('file', file);
-          });
-
-          // Add storage backend information to ensure consistent behavior with single file uploads
-          formData.append('storageBackend', 's3');
-
-          // Note: The unified POST /api/memories endpoint handles user authentication internally
-          // No need to pass userId as it will be determined from the session or onboarding context
-
-          console.log('📤 Uploading folder with storageBackend=s3');
-          const response = await fetch('/api/memories', { method: 'POST', body: formData });
-
-          type FolderResp = {
-            error?: string;
-            userId?: string;
-            successfulUploads?: number;
-            results?: Array<{ memoryId: string; size?: number; checksum_sha256?: string | null }>;
-          };
-          const json = (await response.json()) as FolderResp;
-          data = json;
-
-          if (!response.ok) {
-            throw new Error(json?.error || 'Folder upload failed');
-          }
-        }
-
-        // Best-effort verify first reported memory
-        const appMemoryId = data?.results?.[0]?.memoryId;
-        if (appMemoryId) {
-          await verifyUpload({
-            appMemoryId,
-            database: storage.database,
-            blob_storage: storage.blob_storage,
-            idem: storage.idem,
-            size: data?.results?.[0]?.size || null,
-            checksum_sha256: data?.results?.[0]?.checksum_sha256 || null,
-            remote_id: data?.results?.[0]?.memoryId || appMemoryId,
-          });
-        }
-
-        // Update context with results (onboarding)
-        if (isOnboarding && data?.successfulUploads && data.successfulUploads > 0) {
-          updateUserData({
-            uploadedFileCount: data.successfulUploads,
-            allUserId: data.userId ?? '',
-            memoryId: data.results?.[0]?.memoryId ?? '',
-          });
-
-          if (session) {
-            setCurrentStep('share');
-          } else {
-            setCurrentStep('user-info');
-          }
-        }
-
-        onSuccess?.();
-      } catch (error) {
-        console.error('Folder upload error:', error);
-        toast({
-          variant: 'destructive',
-          title: 'Upload failed',
-          description: error instanceof Error ? error.message : 'Please try again.',
-        });
-        onError?.(error as Error);
       } finally {
         setIsLoading(false);
       }
@@ -436,7 +193,7 @@ export function useFileUpload({ isOnboarding = false, mode = 'directory', onSucc
         await processSingleFile(file, false, userId);
         setIsLoading(false);
       } else if (mode === 'multiple-files') {
-        // Multiple files: use folder upload logic
+        // Multiple files: use multiple files processor
         console.log(`🎯 DASHBOARD MULTIPLE FILES UPLOAD TRIGGERED:`, {
           fileCount: files.length,
           files: Array.from(files).map(f => ({
@@ -449,137 +206,20 @@ export function useFileUpload({ isOnboarding = false, mode = 'directory', onSucc
           isOnboarding,
         });
 
-        // Check file count limit
-        if (!UPLOAD_LIMITS.isFileCountValid(files.length)) {
-          toast({
-            variant: 'destructive',
-            title: 'Too many files',
-            description: UPLOAD_LIMITS.getFileCountErrorMessage(files.length),
-          });
-          return;
-        }
-
-        // Check total size limit
-        const totalSize = Array.from(files).reduce((sum, file) => sum + file.size, 0);
-        if (!UPLOAD_LIMITS.isTotalSizeValid(totalSize)) {
-          toast({
-            variant: 'destructive',
-            title: 'Upload too large',
-            description: UPLOAD_LIMITS.getTotalSizeErrorMessage(totalSize),
-          });
-          return;
-        }
-
         setIsLoading(true);
-
         try {
-          // 1) Request upload storage (MVP mock)
-          const storageResponse = await verifyIntent({
-            preferred: preferences?.blobHosting,
-            databaseHosting: preferences?.databaseHosting,
-            backendHosting: preferences?.backendHosting,
+          await processMultipleFilesService({
+            files: Array.from(files),
+            mode: 'multiple-files',
+            isOnboarding,
+            preferences,
+            onSuccess,
+            onError,
+            updateUserData,
+            setCurrentStep,
+            session,
+            showToast: toast,
           });
-          const storage = storageResponse.uploadStorage;
-
-          let data:
-            | {
-                results?: Array<{
-                  memoryId: string;
-                  size?: number;
-                  checksum_sha256?: string | null;
-                  name?: string;
-                  type?: string;
-                }>;
-                userId?: string;
-                successfulUploads?: number;
-              }
-            | undefined;
-
-          // 2) Route to appropriate upload service based on chosen storage
-          if (storage.blob_storage === 'icp') {
-            // Pre-check Internet Identity authentication before attempting ICP upload
-            const { icpUploadService } = await import('@/services/upload/icp-upload');
-            const isAuthenticated = await icpUploadService.isAuthenticated();
-            if (!isAuthenticated) {
-              throw new Error('Please connect your Internet Identity to upload to ICP');
-            }
-
-            const icpResults = await icpUploadService.uploadFolder(Array.from(files), storage, () => {});
-
-            // Convert ICP results to expected format
-            data = {
-              results: icpResults.map((result, index) => ({
-                memoryId: result.memoryId,
-                size: result.size,
-                name: files[index].name,
-                type: files[index].type,
-                checksum_sha256: result.checksum_sha256,
-              })),
-            };
-          } else {
-            // Upload multiple files using unified POST /api/memories endpoint
-            const formData = new FormData();
-            Array.from(files).forEach(file => {
-              formData.append('file', file);
-            });
-
-            // Note: The unified POST /api/memories endpoint handles user authentication internally
-            // No need to pass userId as it will be determined from the session or onboarding context
-
-            const response = await fetch('/api/memories', { method: 'POST', body: formData });
-
-            type MultipleFilesResp = {
-              error?: string;
-              userId?: string;
-              successfulUploads?: number;
-              results?: Array<{ memoryId: string; size?: number; checksum_sha256?: string | null }>;
-            };
-            const json = (await response.json()) as MultipleFilesResp;
-            data = json;
-
-            if (!response.ok) {
-              throw new Error(json?.error || 'Multiple files upload failed');
-            }
-          }
-
-          // Best-effort verify first reported memory
-          const appMemoryId = data?.results?.[0]?.memoryId;
-          if (appMemoryId) {
-            await verifyUpload({
-              appMemoryId,
-              database: storage.database,
-              blob_storage: storage.blob_storage,
-              idem: storage.idem,
-              size: data?.results?.[0]?.size || null,
-              checksum_sha256: data?.results?.[0]?.checksum_sha256 || null,
-              remote_id: data?.results?.[0]?.memoryId || appMemoryId,
-            });
-          }
-
-          // Update context with results (onboarding)
-          if (isOnboarding && data?.successfulUploads && data.successfulUploads > 0) {
-            updateUserData({
-              uploadedFileCount: data.successfulUploads,
-              allUserId: data.userId ?? '',
-              memoryId: data.results?.[0]?.memoryId ?? '',
-            });
-
-            if (session) {
-              setCurrentStep('share');
-            } else {
-              setCurrentStep('user-info');
-            }
-          }
-
-          onSuccess?.();
-        } catch (error) {
-          console.error('Multiple files upload error:', error);
-          toast({
-            variant: 'destructive',
-            title: 'Upload failed',
-            description: error instanceof Error ? error.message : 'Please try again.',
-          });
-          onError?.(error as Error);
         } finally {
           setIsLoading(false);
         }
