+++ src/services/upload/index.ts
@@ -3,4 +3,6 @@ export * from './upload';
 export * from './intent';
 export * from './verification';
 export * from './file-picker';
+export * from './single-file-processor';
+export * from './multiple-files-processor';
 export { icpUploadService } from './icp-upload';
